+++ src/promptflow-tools/tests/test_common.py
@@ -30,12 +30,10 @@ class TestCommon:
     @pytest.mark.parametrize(
         "function_call, error_message",
         [
-            ({"name": "get_current_weather"}, "must be str, bytes or bytearray"),
-            ("{'name': 'get_current_weather'}", "is an invalid json"),
-            ("get_current_weather", "is an invalid json"),
             ("123", "function_call parameter '123' must be a dict"),
-            ('{"name1": "get_current_weather"}', 'function_call parameter {"name1": "get_current_weather"} must '
-                                                 'contain "name" field'),
+            ({"name1": "get_current_weather"},
+             'function_call parameter {"name1": "get_current_weather"} must '
+             'contain "name" field'),
         ],
     )
     def test_chat_api_invalid_function_call(self, function_call, error_message):
@@ -56,13 +54,23 @@ class TestCommon:
         [
             ("system:\nthis is my function:\ndef hello", [
                 {'role': 'system', 'content': 'this is my function:\ndef hello'}]),
+            ("#system:\nthis is my ##function:\ndef hello", [
+                {'role': 'system', 'content': 'this is my ##function:\ndef hello'}]),
             (" \n system:\nthis is my function:\ndef hello", [
+                {'role': 'system', 'content': 'this is my function:\ndef hello'}]),
+            (" \n # system:\nthis is my function:\ndef hello", [
                 {'role': 'system', 'content': 'this is my function:\ndef hello'}]),
             ("user:\nhi\nassistant:\nanswer\nfunction:\nname:\nn\ncontent:\nc", [
                 {'role': 'user', 'content': 'hi'},
                 {'role': 'assistant', 'content': 'answer'},
                 {'role': 'function', 'name': 'n', 'content': 'c'}]),
+            ("#user :\nhi\n #assistant:\nanswer\n# function:\n##name:\nn\n##content:\nc", [
+                {'role': 'user', 'content': 'hi'},
+                {'role': 'assistant', 'content': 'answer'},
+                {'role': 'function', 'name': 'n', 'content': 'c'}]),
             ("\nsystem:\nfirst\n\nsystem:\nsecond", [
+                {'role': 'system', 'content': 'first'}, {'role': 'system', 'content': 'second'}]),
+            ("\n#system:\nfirst\n\n#system:\nsecond", [
                 {'role': 'system', 'content': 'first'}, {'role': 'system', 'content': 'second'}])
         ]
     )