+++ src/promptflow/promptflow/_cli/_pf/_init_entry_generators.py
@@ -6,6 +6,7 @@ import inspect
 import logging
 import shutil
 from abc import ABC, abstractmethod
+from ast import literal_eval
 from enum import Enum
 from pathlib import Path
 
@@ -15,8 +16,10 @@ from promptflow._sdk._constants import LOGGER_NAME
 
 logger = logging.getLogger(LOGGER_NAME)
 TEMPLATE_PATH = Path(__file__).parent.parent / "data" / "entry_flow"
+CHAT_FLOW_TEMPLATE_PATH = Path(__file__).parent.parent / "data" / "chat_flow" / "template"
 TOOL_TEMPLATE_PATH = Path(__file__).parent.parent / "data" / "package_tool"
 EXTRA_FILES_MAPPING = {"requirements.txt": "requirements_txt", ".gitignore": "gitignore"}
+SERVE_TEMPLATE_PATH = Path(__file__).resolve().parent.parent.parent / "_sdk" / "data" / "executable"
 
 
 class BaseGenerator(ABC):
@@ -217,6 +220,78 @@ class FlowMetaYamlGenerator(BaseGenerator):
         return ["flow_name"]
 
 
+class StreamlitFileGenerator(BaseGenerator):
+    def __init__(self, flow_name, flow_dag_path):
+        self.flow_name = flow_name
+        self.flow_dag_path = Path(flow_dag_path)
+
+    @property
+    def flow_inputs(self):
+        from promptflow.contracts.flow import Flow as ExecutableFlow
+
+        executable = ExecutableFlow.from_yaml(
+            flow_file=Path(self.flow_dag_path.name), working_dir=self.flow_dag_path.parent
+        )
+        return {flow_input: (value.default, value.type.value) for flow_input, value in executable.inputs.items()}
+
+    @property
+    def flow_inputs_params(self):
+        flow_inputs_params = ["=".join([flow_input, flow_input]) for flow_input, _ in self.flow_inputs.items()]
+        return ",".join(flow_inputs_params)
+
+    @property
+    def tpl_file(self):
+        return SERVE_TEMPLATE_PATH / "main.py.jinja2"
+
+    @property
+    def flow_path(self):
+        return self.flow_dag_path.as_posix()
+
+    @property
+    def entry_template_keys(self):
+        return ["flow_name", "flow_inputs", "flow_inputs_params", "flow_path"]
+
+
+class ChatFlowDAGGenerator(BaseGenerator):
+    def __init__(self, connection, deployment):
+        self.connection = connection
+        self.deployment = deployment
+
+    @property
+    def tpl_file(self):
+        return CHAT_FLOW_TEMPLATE_PATH / "flow.dag.yaml.jinja2"
+
+    @property
+    def entry_template_keys(self):
+        return ["connection", "deployment"]
+
+
+class AzureOpenAIConnectionGenerator(BaseGenerator):
+    def __init__(self, connection):
+        self.connection = connection
+
+    @property
+    def tpl_file(self):
+        return CHAT_FLOW_TEMPLATE_PATH / "azure_openai.yaml.jinja2"
+
+    @property
+    def entry_template_keys(self):
+        return ["connection"]
+
+
+class OpenAIConnectionGenerator(BaseGenerator):
+    def __init__(self, connection):
+        self.connection = connection
+
+    @property
+    def tpl_file(self):
+        return CHAT_FLOW_TEMPLATE_PATH / "openai.yaml.jinja2"
+
+    @property
+    def entry_template_keys(self):
+        return ["connection"]
+
+
 def copy_extra_files(flow_path, extra_files):
     for file_name in extra_files:
         extra_file_path = (
@@ -229,8 +304,23 @@ def copy_extra_files(flow_path, extra_files):
 
 
 class ToolPackageGenerator(BaseGenerator):
-    def __init__(self, tool_name):
+    def __init__(self, tool_name, icon=None, extra_info=None):
         self.tool_name = tool_name
+        self._extra_info = extra_info
+        self.icon = icon
+
+    @property
+    def extra_info(self):
+        if self._extra_info:
+            extra_info = {}
+            for k, v in self._extra_info.items():
+                try:
+                    extra_info[k] = literal_eval(v)
+                except Exception:
+                    extra_info[k] = repr(v)
+            return extra_info
+        else:
+            return {}
 
     @property
     def tpl_file(self):
@@ -238,7 +328,20 @@ class ToolPackageGenerator(BaseGenerator):
 
     @property
     def entry_template_keys(self):
-        return ["tool_name"]
+        return ["tool_name", "extra_info", "icon"]
+
+
+class ManifestGenerator(BaseGenerator):
+    def __init__(self, package_name):
+        self.package_name = package_name
+
+    @property
+    def tpl_file(self):
+        return TOOL_TEMPLATE_PATH / "MANIFEST.in.jinja2"
+
+    @property
+    def entry_template_keys(self):
+        return ["package_name"]
 
 
 class SetupGenerator(BaseGenerator):