+++ src/promptflow/tests/executor/unittests/_core/test_tools_manager.py
@@ -167,6 +167,7 @@ class TestToolsManager:
         }
 
         expected_template = {
+            "$schema": "https://azuremlschemas.azureedge.net/promptflow/latest/CustomStrongTypeConnection.schema.json",
             "name": "to_replace_with_connection_name",
             "type": "custom",
             "custom_type": "MyFirstConnection",
@@ -190,12 +191,23 @@ class TestToolsManager:
             package: test-custom-tools
             package_version: 0.0.2
             configs:
-              api_url: "This is a fake api url."  # String, The api url.
+              api_url: "This is a fake api url."  # String type. The api url.
             secrets:      # must-have
-              api_key: "to_replace_with_api_key"  # String, The api key.
+              api_key: "to_replace_with_api_key"  # String type. The api key.
             """
 
         content = templates["my_tool_package.tools.my_tool_with_custom_strong_type_connection.MyCustomConnection"]
         expected_template_str = textwrap.dedent(expected_template)
+        assert expected_template_str in content
 
-        assert content in expected_template_str
+    # TODO: enable this test after new my_tool_package is released
+    @pytest.mark.skip("Will enable this test after new my_tool_package is released")
+    def test_gen_dynamic_list(self):
+        from promptflow._sdk._utils import _gen_dynamic_list
+        func_path = "my_tool_package.tools.tool_with_dynamic_list_input.my_list_func"
+        func_kwargs = {"prefix": "My"}
+        result = _gen_dynamic_list({
+            "func_path": func_path, "func_kwargs": func_kwargs})
+        assert len(result) == 10
+
+    # TODO: add test for gen_dynamic_list with ws_triple.