+++ src/promptflow/tests/executor/unittests/executor/test_tool_resolver.py
@@ -1,3 +1,4 @@
+import re
 import sys
 from pathlib import Path
 
@@ -20,7 +21,7 @@ from promptflow.executor._errors import (
 )
 from promptflow.executor._tool_resolver import ResolvedTool, ToolResolver
 
-from ...utils import FLOW_ROOT
+from ...utils import DATA_ROOT, FLOW_ROOT
 
 TEST_ROOT = Path(__file__).parent.parent.parent
 REQUESTS_PATH = TEST_ROOT / "test_configs/executor_api_requests"
@@ -289,7 +290,7 @@ class TestToolResolver:
         connections = {"conn_name": {"type": "AzureOpenAIConnection", "value": {"api_key": "mock", "api_base": "mock"}}}
         tool_resolver = ToolResolver(working_dir=None, connections=connections)
         tool_resolver._tool_loader = tool_loader
-        mocker.patch.object(tool_resolver, "_load_source_content", return_value="{{text}}")
+        mocker.patch.object(tool_resolver, "_load_source_content", return_value="{{text}}![image]({{image}})")
 
         node = Node(
             name="mock",
@@ -297,14 +298,17 @@ class TestToolResolver:
             inputs={
                 "conn": InputAssignment(value="conn_name", value_type=InputValueType.LITERAL),
                 "text": InputAssignment(value="Hello World!", value_type=InputValueType.LITERAL),
+                "image": InputAssignment(value=str(DATA_ROOT / "test_image.jpg"), value_type=InputValueType.LITERAL),
             },
             connection="conn_name",
             provider="mock",
         )
         resolved_tool = tool_resolver._resolve_llm_node(node, convert_input_types=True)
-        assert len(resolved_tool.node.inputs) == 1
+        assert len(resolved_tool.node.inputs) == 2
         kwargs = {k: v.value for k, v in resolved_tool.node.inputs.items()}
-        assert resolved_tool.callable(**kwargs) == "Hello World!"
+        pattern = re.compile(r"^Hello World!!\[image\]\(Image\([a-z0-9]{8}\)\)$")
+        prompt = resolved_tool.callable(**kwargs)
+        assert re.match(pattern, prompt)
 
     def test_resolve_script_node(self, mocker):
         def mock_python_func(conn: AzureOpenAIConnection, prompt: PromptTemplate, **kwargs):
@@ -413,9 +417,12 @@ class TestToolResolver:
         node = mocker.Mock(name="node", tool=None, inputs={})
         node.type = ToolType.PYTHON
         node.source = mocker.Mock(type=ToolSourceType.Code)
+        tool = Tool(name="tool", type="python", inputs={"conn": InputDefinition(type=["CustomConnection"])})
         m = sys.modules[__name__]
         v = InputAssignment(value="conn_name", value_type=InputValueType.LITERAL)
-        actual = tool_resolver._convert_to_custom_strong_type_connection_value("conn_name", v, node, conn_types, m)
+        actual = tool_resolver._convert_to_custom_strong_type_connection_value(
+            "conn_name", v, node, tool, conn_types, m
+        )
         assert isinstance(actual, expected_type)
         assert actual.api_base == "mock"