+++ src/promptflow/tests/sdk_cli_test/e2etests/test_cli.py
@@ -1,6 +1,7 @@
 import contextlib
-import io
+import importlib
 import importlib.util
+import io
 import json
 import logging
 import os
@@ -13,6 +14,7 @@ from pathlib import Path
 from tempfile import mkdtemp
 from unittest.mock import patch
 
+import mock
 import pytest
 import yaml
 
@@ -534,7 +536,7 @@ class TestCli:
                 assert details["node_runs"][0]["logs"]["stdout"]
 
         env = {"API_BASE": "${azure_open_ai_connection.api_base}"}
-        SubmitterHelper.resolve_environment_variables(env)
+        SubmitterHelper.resolve_environment_variables(env, local_client)
         run_pf_command(
             "flow",
             "test",
@@ -762,6 +764,94 @@ class TestCli:
             chat_prompt_node = next(filter(lambda item: item["name"] == "chat_prompt", flow_dict["nodes"]))
             assert "chat_history" in chat_prompt_node["inputs"]
             assert "customer_info" in chat_prompt_node["inputs"]
+
+    def test_flow_init_with_connection_and_deployment(self):
+        def check_connection_and_deployment(flow_folder, connection, deployment):
+            with open(Path(flow_folder) / "flow.dag.yaml", "r") as f:
+                flow_dict = yaml.safe_load(f)
+                assert flow_dict["nodes"][0]["inputs"]["deployment_name"] == deployment
+                assert flow_dict["nodes"][0]["connection"] == connection
+
+        temp_dir = mkdtemp()
+        with _change_working_dir(temp_dir):
+            flow_name = "chat_flow"
+            flow_folder = Path(temp_dir) / flow_name
+            # When configure local connection provider, init chat flow without connection and deployment.
+            run_pf_command(
+                "flow",
+                "init",
+                "--flow",
+                flow_name,
+                "--type",
+                "chat",
+            )
+            # Assert connection files created
+            assert (flow_folder / "azure_openai.yaml").exists()
+            assert (flow_folder / "openai.yaml").exists()
+
+            # When configure local connection provider, init chat flow with connection and deployment.
+            connection = "connection_name"
+            deployment = "deployment_name"
+            run_pf_command(
+                "flow",
+                "init",
+                "--flow",
+                flow_name,
+                "--type",
+                "chat",
+                "--connection",
+                connection,
+                "--deployment",
+                deployment,
+                "--yes",
+            )
+            # Assert connection files created and the connection/deployment is set in flow.dag.yaml
+            check_connection_and_deployment(flow_folder, connection=connection, deployment=deployment)
+            connection_files = [flow_folder / "azure_openai.yaml", flow_folder / "openai.yaml"]
+            for file in connection_files:
+                assert file.exists()
+                with open(file, "r") as f:
+                    connection_dict = yaml.safe_load(f)
+                    assert connection_dict["name"] == connection
+
+            shutil.rmtree(flow_folder)
+            target = "promptflow._sdk._pf_client.Configuration.get_connection_provider"
+            with mock.patch(target) as mocked:
+                mocked.return_value = "azureml:xx"
+                # When configure azure connection provider, init chat flow without connection and deployment.
+                run_pf_command(
+                    "flow",
+                    "init",
+                    "--flow",
+                    flow_name,
+                    "--type",
+                    "chat",
+                    "--yes",
+                )
+                # Assert connection files not created.
+                assert not (flow_folder / "azure_openai.yaml").exists()
+                assert not (flow_folder / "openai.yaml").exists()
+
+                # When configure azure connection provider, init chat flow with connection and deployment.
+                connection = "connection_name"
+                deployment = "deployment_name"
+                run_pf_command(
+                    "flow",
+                    "init",
+                    "--flow",
+                    flow_name,
+                    "--type",
+                    "chat",
+                    "--connection",
+                    connection,
+                    "--deployment",
+                    deployment,
+                    "--yes",
+                )
+                # Assert connection files not created and the connection/deployment is set in flow.dag.yaml
+                check_connection_and_deployment(flow_folder, connection=connection, deployment=deployment)
+                assert not (flow_folder / "azure_openai.yaml").exists()
+                assert not (flow_folder / "openai.yaml").exists()
 
     def test_flow_chat(self, monkeypatch, capsys):
         chat_list = ["hi", "what is chat gpt?"]
@@ -1224,7 +1314,8 @@ class TestCli:
             assert (package_folder / "README.md").exists()
 
             spec = importlib.util.spec_from_file_location(
-                f"{package_name}.utils", package_folder / package_name / "utils.py")
+                f"{package_name}.utils", package_folder / package_name / "utils.py"
+            )
             utils = importlib.util.module_from_spec(spec)
             spec.loader.exec_module(utils)
 
@@ -1254,6 +1345,86 @@ class TestCli:
             outerr = capsys.readouterr()
             assert f"The tool name {invalid_tool_name} is a invalid identifier." in outerr.out
 
+            # Test init package tool with extra info
+            package_name = "tool_with_extra_info"
+            package_folder = Path(temp_dir) / package_name
+            icon_path = Path(DATAS_DIR) / "logo.jpg"
+            category = "test_category"
+            tags = {'tag1': 'value1', 'tag2': 'value2'}
+            run_pf_command(
+                "tool",
+                "init",
+                "--package",
+                package_name,
+                "--tool",
+                func_name,
+                "--set",
+                f"icon={icon_path.absolute()}",
+                f"category={category}",
+                f"tags={tags}",
+                cwd=temp_dir
+            )
+            spec = importlib.util.spec_from_file_location(
+                f"{package_name}.utils", package_folder / package_name / "utils.py")
+            utils = importlib.util.module_from_spec(spec)
+            spec.loader.exec_module(utils)
+
+            assert hasattr(utils, "list_package_tools")
+            tools_meta = utils.list_package_tools()
+            meta = tools_meta[f"{package_name}.{func_name}.{func_name}"]
+            assert meta["category"] == category
+            assert meta["tags"] == tags
+            assert meta["icon"].startswith("data:image")
+
+            # icon doesn't exist
+            with pytest.raises(SystemExit):
+                run_pf_command(
+                    "tool",
+                    "init",
+                    "--package",
+                    package_name,
+                    "--tool",
+                    func_name,
+                    "--set",
+                    "icon=invalid_icon_path",
+                    cwd=temp_dir
+                )
+            outerr = capsys.readouterr()
+            assert "Cannot find the icon path" in outerr.out
+
+    def test_tool_list(self, capsys):
+        # List package tools in environment
+        run_pf_command("tool", "list")
+        outerr = capsys.readouterr()
+        tools_dict = json.loads(outerr.out)
+        package_tool_name = "promptflow.tools.embedding.embedding"
+        assert package_tool_name in tools_dict["package"]
+
+        # List flow tools and package tools
+        run_pf_command("tool", "list", "--flow", f"{FLOWS_DIR}/chat_flow")
+        outerr = capsys.readouterr()
+        tools_dict = json.loads(outerr.out)
+        expect_flow_tools = {
+            "chat.jinja2": {
+                "type": "llm",
+                "inputs": {"chat_history": {"type": ["string"]}, "question": {"type": ["string"]}},
+                "source": "chat.jinja2",
+            },
+            "show_answer.py": {
+                "type": "python",
+                "inputs": {"chat_answer": {"type": ["string"]}},
+                "source": "show_answer.py",
+                "function": "show_answer",
+            },
+        }
+        assert tools_dict["code"] == expect_flow_tools
+        assert package_tool_name in tools_dict["package"]
+
+        # Invalid flow parameter
+        with pytest.raises(Exception) as e:
+            run_pf_command("tool", "list", "--flow", "invalid_flow_folder")
+        assert "invalid_flow_folder does not exist" in e.value.args[0]
+
     def test_chat_flow_with_conditional(self, monkeypatch, capsys):
         chat_list = ["1", "2"]
 
@@ -1265,14 +1436,21 @@ class TestCli:
 
         monkeypatch.setattr("builtins.input", mock_input)
         run_pf_command(
-            "flow",
-            "test",
-            "--flow",
-            f"{FLOWS_DIR}/conditional_chat_flow_with_skip",
-            "--interactive",
-            "--verbose"
+            "flow", "test", "--flow", f"{FLOWS_DIR}/conditional_chat_flow_with_skip", "--interactive", "--verbose"
         )
         output_path = Path(FLOWS_DIR) / "conditional_chat_flow_with_skip" / ".promptflow" / "chat.output.json"
         assert output_path.exists()
         detail_path = Path(FLOWS_DIR) / "conditional_chat_flow_with_skip" / ".promptflow" / "chat.detail.json"
         assert detail_path.exists()
+
+    def test_flow_test_with_image_input_and_output(self):
+        run_pf_command(
+            "flow",
+            "test",
+            "--flow",
+            f"{FLOWS_DIR}/python_tool_with_simple_image",
+        )
+        output_path = Path(FLOWS_DIR) / "python_tool_with_simple_image" / ".promptflow" / "output"
+        assert output_path.exists()
+        image_path = Path(FLOWS_DIR) / "python_tool_with_simple_image" / ".promptflow" / "intermediate"
+        assert image_path.exists()