+++ src/promptflow/tests/test_configs/flows/python_tool_with_composite_image/passthrough_dict.py
@@ -1,0 +1,6 @@
+from promptflow import tool
+
+
+@tool
+def passthrough_dict(image_list: list,image_dict: dict) -> list:
+    return image_dict