+++ src/promptflow/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Features Added
 - [Executor] Add average execution time and estimated execution time to batch run logs
+- [SDK/CLI] Support `pfazure run archive/restore/update`.
+- [SDK/CLI] Support custom strong type connection.
+- [SDK/CLI] Enable telemetry and won't collect by default, use `pf config set cli.telemetry_enabled=true` to opt in.
 
 ### Bugs Fixed
 - **pf config set**:
@@ -33,7 +36,7 @@
 - **pf flow validate**: support validate flow
 - **pf config set**: support set user-level promptflow config.
   - Support workspace connection provider, usage: `pf config set connection.provider=azureml:/subscriptions/<subscription_id>/resourceGroups/<resource_group>/providers/Microsoft.MachineLearningServices/workspaces/<workspace_name>`
-- Support override openai connection's model when submitting a flow. For example: `pf run create --flow ./ --data ./data.jsonl --connection llm.model=xxx`
+- Support override openai connection's model when submitting a flow. For example: `pf run create --flow ./ --data ./data.jsonl --connection llm.model=xxx --column-mapping url='${data.url}'`
 
 ### Bugs Fixed
 - [Flow build] Fix flow build file name and environment variable name when connection name contains space.